% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BassSIRstan-package.R
\docType{package}
\name{BassSIRstan-package}
\alias{BassSIRstan-package}
\alias{BassSIRstan}
\title{The 'BassSIRstan' package.}
\description{
Infectious disease modelling with Bass-SIR model
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
