% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_two_step}
\alias{fit_two_step}
\title{Fit BassSIR model to data with a two step approach}
\usage{
fit_two_step(
  sel,
  loc = "X",
  r_iso = 1/6.1,
  n_iter_1 = 10000,
  n_iter_2 = 1e+05,
  n_collect = 5000,
  n_chain = 3,
  thres = 1.02,
  ...
)
}
\arguments{
\item{sel}{selected data set}

\item{loc}{location of interest}

\item{r_iso}{isolation rate (invasion of isolation delay)}

\item{n_iter_1}{number of iterations for the absorbing state model}

\item{n_iter_2}{number of iterations for the BassSIR model}

\item{n_collect}{number of iterations being collected as output}

\item{n_chain}{number of chains}

\item{thres}{threshold of Rhat for testing convergance}

\item{...}{argumens for rstan::sampling(...)}
}
\value{

}
\description{
Fit BassSIR model to data with a two step approach
}
\examples{
f_bass_sir = fit_two_step(tail(n_by_country$US, 20), "US", refresh = 0)
summary(f_bass_sir)

}
