% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.estBassSIR}
\alias{fitted.estBassSIR}
\title{Fitted values of BassSIR models}
\usage{
\method{fitted}{estBassSIR}(est)
}
\arguments{
\item{est}{estimated model}
}
\value{

}
\description{
Fitted values of BassSIR models
}
\examples{
f_bass_sir = fit_two_step(tail(n_by_country$US, 20), "US", refresh = 0)
fitted(f_bass_sir)

}
