% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_absorb_data}
\alias{get_absorb_data}
\alias{get_two_step_data}
\title{Reform data to model inputs}
\usage{
get_absorb_data(nc)

get_two_step_data(nc, f_abs)
}
\arguments{
\item{nc}{notification counts generated by as_bass_data}

\item{f_abs}{results from absorb model if two-step approach applied}
}
\value{

}
\description{
Reform data to model inputs
}
\examples{
nc <- as_bass_data(tail(n_by_country$US, 20), "US", status = "Cumulative")
input <- get_absorb_data(nc)

}
