% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{n_by_country}
\alias{n_by_country}
\title{Notication data of COVID-19}
\format{
A list of data frames with columns:
\describe{
\item{Location}{A country or region}
\item{Date}{Date of the record}
\item{Confirmed}{Confirmed cases}
\item{Cured}{Recovered}
\item{Dead}{Deaths}
}
}
\source{
Fetched and rearranged from CSSE data \url{https://github.com/CSSEGISandData/COVID-19/tree/master/csse_covid_19_data}
}
\usage{
n_by_country
}
\description{
Notication data of COVID-19
}
\examples{
\dontrun{
 n_by_country
}
}
\keyword{datasets}
