% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_absorb}
\alias{sample_absorb}
\alias{sample_bass}
\title{Sample from stan models}
\usage{
sample_absorb(
  d,
  r_iso,
  pars = c("r_death", "r_rec", "I", "cdg", "hi_death", "hi_rec"),
  ...
)

sample_bass(
  d,
  r_iso,
  r_out = d$r_out,
  pars = c("beta", "kappa", "m", "I_hat", "R0", "Rt", "PrEx"),
  ...
)
}
\arguments{
\item{d}{data to be fitted}

\item{r_iso}{rate of isolation}

\item{pars}{parameters for outputing}

\item{...}{arguments from rstan::sampling}

\item{r_out}{rate of recovery and death}
}
\value{
mcmc results from rstan
}
\description{
Sample from stan models
}
