% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary.estBassSIR}
\alias{summary.estBassSIR}
\alias{print.summaryEstBassSIR}
\title{Summarise a fitted Bass SIR model}
\usage{
\method{summary}{estBassSIR}(est)

\method{print}{summaryEstBassSIR}(est)
}
\arguments{
\item{est}{a model with fitted parameters}
}
\value{

}
\description{
Summarise a fitted Bass SIR model
}
\examples{
f_bass_sir = fit_two_step(tail(n_by_country$US, 20), "US", refresh = 0)
summary(f_bass_sir)

}
